waithandle renamed to match coreclr, Release Mutex Core removed
