+++ src/System.Private.CoreLib/shared/System/Threading/Mutex.cs
@@ -58,36 +58,5 @@ namespace System.Threading
 
         public static bool TryOpenExisting(string name, out Mutex result) =>
             OpenExistingWorker(name, out result) == OpenExistingResult.Success;
-
-        // Note: To call ReleaseMutex, you must have an ACL granting you
-        // MUTEX_MODIFY_STATE rights (0x0001). The other interesting value
-        // in a Mutex's ACL is MUTEX_ALL_ACCESS (0x1F0001).
-        public void ReleaseMutex()
-        {
-#if CORECLR
-            if (!Interop.Kernel32.ReleaseMutex(safeWaitHandle))
-            {
-                throw new ApplicationException(SR.Arg_SynchronizationLockException);
-            }
-#else
-            // The field value is modifiable via the public <see cref="WaitHandle.SafeWaitHandle"/> property, save it locally
-            // to ensure that one instance is used in all places in this method
-            SafeWaitHandle waitHandle = _waitHandle;
-            if (waitHandle == null)
-            {
-                ThrowInvalidHandleException();
-            }
-
-            waitHandle.DangerousAddRef();
-            try
-            {
-                ReleaseMutexCore(waitHandle);
-            }
-            finally
-            {
-                waitHandle.DangerousRelease();
-            }
-#endif
-        }
     }
 }
