+++ src/System.Private.CoreLib/src/System/Threading/Semaphore.cs
@@ -95,7 +95,7 @@ namespace System.Threading
         {
             // The field value is modifiable via the public <see cref="WaitHandle.SafeWaitHandle"/> property, save it locally
             // to ensure that one instance is used in all places in this method
-            SafeWaitHandle waitHandle = _waitHandle;
+            SafeWaitHandle waitHandle = safeWaitHandle;
             if (waitHandle == null)
             {
                 ThrowInvalidHandleException();
